moving getters functions
