+++ src/views/catalog/customize/components/CustomizeForm.tsx
@@ -5,7 +5,7 @@ import { V1Template } from '@kubevirt-ui/kubevirt-api/console';
 import { useKubevirtTranslation } from '@kubevirt-utils/hooks/useKubevirtTranslation';
 import { ActionGroup, Button, ExpandableSection, Form } from '@patternfly/react-core';
 
-import { getTemplateVirtualMachineObject } from '../../utils/vm-template-source/utils';
+import { getTemplateVirtualMachineObject } from '../../utils/templateGetters';
 import { useWizardVMContext } from '../../utils/WizardVMContext';
 import {
   buildFields,
