+++ src/views/catalog/templatescatalog/components/TemplatesCatalogDrawer/TemplatesCatalogDrawer.tsx
@@ -4,7 +4,7 @@ import { V1Template } from '@kubevirt-ui/kubevirt-api/console';
 import { CatalogItemHeader } from '@patternfly/react-catalog-view-extension';
 import { Modal } from '@patternfly/react-core';
 
-import { getTemplateName } from '../../utils/helpers';
+import { getTemplateName } from '../../../utils/templateGetters';
 import { getTemplateOSIcon } from '../../utils/os-icons';
 
 import { TemplatesCatalogDrawerFooter } from './TemplatesCatalogDrawerFooter';
