+++ src/views/catalog/utils/annotations.ts
@@ -3,4 +3,8 @@ export const ANNOTATIONS = Object.freeze({
   providerDisplayName: 'openshift.io/provider-display-name',
   documentationURL: 'openshift.io/documentation-url',
   supportURL: 'openshift.io/support-url',
+  description: 'description',
+  os: 'vm.kubevirt.io/os',
+  osTemplate: 'vm.kubevirt.io/template',
+  supportLevel: 'template.kubevirt.io/provider-support-level',
 });
