+++ src/views/catalog/utils/constants.ts
@@ -4,11 +4,6 @@ export const TEMPLATE_TYPE_VM = 'vm';
 export const TEMPLATE_TYPE_BASE = 'base';
 export const TEMPLATE_FLAVOR_LABEL = 'flavor.template.kubevirt.io';
 export const TEMPLATE_WORKLOAD_LABEL = 'workload.template.kubevirt.io';
-export const TEMPLATE_SUPPORT_LEVEL_ANNOTATION = 'template.kubevirt.io/provider-support-level';
-export const TEMPLATE_OS_TEMPLATE_ANNOTATION = 'vm.kubevirt.io/template';
-export const TEMPLATE_OS_ANNOTATION = 'vm.kubevirt.io/os';
-export const TEMPLATE_PROVIDER_NAME_ANNOTATION = 'openshift.io/provider-display-name';
-export const TEMPLATE_DESCRIPTION_ANNOTATION = 'description';
 export const TEMPLATE_BASE_IMAGE_NAME_PARAMETER = 'SRC_PVC_NAME';
 export const TEMPLATE_BASE_IMAGE_NAMESPACE_PARAMETER = 'SRC_PVC_NAMESPACE';
 export const TEMPLATE_DATA_SOURCE_NAME_PARAMETER = 'DATA_SOURCE_NAME';
